Crate trillium_proxy

http reverse and forward proxy trillium handler

Modules

upstream

Upstream selectors

Structs

Client

A client contains a Config and an optional connection pool and builds conns.

ForwardProxyConnect

trillium handler to implement Connect proxying

Proxy

the proxy handler

Url

A parsed URL record.

Traits

Connector

Interface for runtime and tls adapters for the trillium client

Functions

proxy

constructs a new Proxy. alias of Proxy::new