Struct trillium_proxy::Proxy

pub struct Proxy<U> { /* private fields */ }

the proxy handler

Implementations

impl<U: UpstreamSelector> Proxy<U>

pub fn new<I>(client: impl Into<Client>, upstream: I) -> Self
where I: IntoUpstreamSelector<UpstreamSelector = U>,

construct a new proxy handler that sends all requests to the upstream provided

use trillium_smol::ClientConfig;
use trillium_proxy::Proxy;

let proxy = Proxy::new(ClientConfig::default(), "http://docs.trillium.rs/trillium_proxy");

pub fn proxy_not_found(self) -> Self

chainable constructor to set the 404 Not Found handling behavior. By default, this proxy will pass through the trillium Conn unmodified if the proxy response is a 404 not found, allowing it to be chained in a tuple handler. To modify this behavior, call proxy_not_found, and the full 404 response will be forwarded. The Conn will be halted unless Proxy::without_halting was configured

let proxy = Proxy::new(ClientConfig::default(), "http://trillium.rs")
    .proxy_not_found();

pub fn without_halting(self) -> Self

The default behavior for this handler is to halt the conn on any response other than a 404. If Proxy::proxy_not_found has been configured, the default behavior for all response statuses is to halt the trillium conn. To change this behavior, call without_halting when constructing the proxy, and it will not halt the conn. This is useful when passing the proxy reply through trillium_html_rewriter.

let proxy = Proxy::new(ClientConfig::default(), "http://trillium.rs")
    .without_halting();

pub fn with_via_pseudonym( self, via_pseudonym: impl Into<Cow<'static, str>> ) -> Self

populate the pseudonym for a Via header. If no pseudonym is provided, no via header will be inserted.

pub fn with_websocket_upgrades(self) -> Self

Allow websockets to be proxied

This is not currently the default, but that may change at some (semver-minor) point in the future