Module trillium_proxy::upstream

Upstream selectors

Structs

ConnectionCounting

an upstream selector that attempts to send requests to the upstream with the fewest open connections.

ForwardProxy

an upstream selector for forward proxy

RandomSelector

a random upstream selector

RoundRobin

an upstream selector that dispatches to upstreams in the order that they’re specified, wrapping

Traits

IntoUpstreamSelector

represents something that can be used as an upstream selector

UpstreamSelector

a trait for selecting the correct upstream